+++ packages/logger/lib/index.js
@@ -1,0 +1,5 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+// @kaapi/logger
+// logger
+//# sourceMappingURL=index.js.map